requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
django==5.2.3
django-crispy-forms==2.4
django-filter==25.1
django-import-export==4.3.8
pip==25.1.1
sqlparse==0.5.3
tablib==3.8.0
xhtml2pdf==0.2.15
